feat: Add Family Yearly Report view
Introduces a new view to display the tithe contributions of a specific family for an entire year.
This includes fetching and displaying monthly tithe data per category and calculating yearly totals.
A "Report" button has been added to each family row in the TitheTable to navigate to this new view.
The `FamilyYearlyTitheData` type has been added to support this feature.